cloud_functions/sentiment/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-language
google-cloud-firestore